Fix import block & Change if to try & Create local module
load_external_module to use deoplete.util
Remove get_var helper function
Change if...else to try...except
Create 'deoplete_jedi' local module
